Add lax ssl certificate requirements
It happens that Heroku's Redis solution uses a secure connection, which
is then rejected as it has a self-signed certificate. To fix this,
Heroku has provided a guide on disabling the certificate check:

https://devcenter.heroku.com/articles/connecting-heroku-redis#connecting-in-django

We also need to bypass the check for the workers, which required an
update of channels-redis to 4.2.0.